Database

+++ db/migrate/20170725062136_users.rb
@@ -2,7 +2,7 @@ class Users < ActiveRecord::Migration[5.1]
   def change
   	create_table(:users) do |t|
   		t.column(:name, :varchar)
-  		t.column(:Telephone, :integer)
+  		t.column(:Telephone, :bigint)
   		t.column(:Email, :varchar)
 
   		t.timestamps()
